Update roles on roletemplate changes
User rbac controllers create and update clusterRoles and roles based on roleTemplate.
The clusterRoles get updated when a roleTemplate changes. But the namespaced roles
aren't getting updated.
So "get/list/watch" permissions get removed from the roletemplate and clusterRole
"cluster-member" on upgrade, but not from the role "cluster-member" created in the
cluster's namespace.
This commit updates roles too on based on roleTemplates