Restricted admin role changes
One clusterRole for all access to all management CRDs, bound to any restricted admins
in every cluster except local.
One clusterRole for all access to all project CRDs, bound to any restricted admins in each
project except local cluster's projects.
A clusterRole with * permissions on all clusters except the local cluster, with a clusterRoleBinding
for each restricted admin.
The catalog.cattle.io CRDs are live in the clusters they are created for so we don't need to give restricted-admin
role permissions for this